When (re-)generating the CSV file use minimal quoting
(Quote only when the field contains `,` `"` or a new line, not when it
contains a space, but not the other characters.)

This is a rather hack-ish solution. It's also the first PHP code that
I've ever written (other than fiddling with mediawiki and wordpress
templates several years ago), so it may well not be very pretty
code. I haven't seen a noticeable slow-down when running `composer
index` on the anki-ultimate-geography deck, though.

I decided to write with fputcsv to a temporary buffer and then patch
this output, to avoid writing my own CSV parser and possibly introduce
my own bugs.